+++ Sphinx/ChangeLog.rst
@@ -2,6 +2,7 @@ ChangeLog
 =========
 * September 1st, 2014
 
+  - Added ``plot_scatter()`` function in ``simpleplot`` module.
   - Updated documentation.
 
 * 01.06.03 --- July 24, 2014